выполняет д/з модуляция
